#11 general optimization

+++ include/main.h
@@ -14,18 +14,24 @@
 // pin that is connected to rst
 #define RST PB0
 
-unsigned long int c_time = 0;
-unsigned long int l_time_50 = 0;		// last time a 50Hz event occured
-unsigned long int l_time_10 = 0;		// last time a 10Hz event occured
-unsigned long int l_time_5 = 0;			// last time a 5Hz event occured
-unsigned long int l_time_1 = 0;			// last time a 1Hz event occured
+
+// TODO: replace all of this with a timer class
+uint32_t c_time = 0;
+uint32_t l_time_50 = 0;		    // last time a 50Hz event occured
+uint32_t l_time_10 = 0;		    // last time a 10Hz event occured
+uint32_t l_time_5 = 0;			// last time a 5Hz event occured
+uint32_t l_time_1 = 0;			// last time a 1Hz event occured
+
 
 // constants for each refresh rate
-unsigned int fifty_hz = 20;
-unsigned int ten_hz = 100;
-unsigned int five_hz = 200;
-unsigned int one_hz = 1000;
+uint16_t fifty_hz = 20;
+uint16_t ten_hz = 100;
+uint16_t five_hz = 200;
+uint16_t one_hz = 1000;
+// end todo
 
+
+// TODO: replace all of this with a debounce class
 // debounce vars
 uint32_t diag_fuel_deb = 0;
 uint32_t tr_deb = 0;		// debounce var for top right btn
@@ -33,6 +39,8 @@ uint32_t launch_rpm_plus_deb = 0;
 uint32_t launch_rpm_minus_deb = 0;
 uint32_t launch_thresh_plus_deb = 0;
 uint32_t launch_thresh_minus_deb = 0;
+// end todo
+
 
 // status vars
 bool led_err_st = 0;
